\name{coef.CARBayesST}
\alias{coef.CARBayesST}


\title{
    Extract the regression coefficients from a model.
}
\description{
    This function takes a CARBayesST object and returns the vector of estimated
    regression coefficients based on posterior medians.
}

\usage{
    \method{coef}{CARBayesST}(object, \dots)
}



\arguments{
    \item{object}{
        A CARBayesST fitted model object.
    }
    \item{\dots}{
        Ignored.
    }
}
\author{
    Duncan Lee
}
