\name{fitted.CARBayesST}
\alias{fitted.CARBayesST}


\title{
    Extract the fitted values from a model.
}
\description{
    This function takes a CARBayesST object and returns the vector of fitted
    values (posterior means).
}

\usage{
    \method{fitted}{CARBayesST}(object, \dots)
}



\arguments{
    \item{object}{
        A CARBayesST fitted model object.
    }
    \item{\dots}{
        Ignored.
    }
}
\author{
    Duncan Lee
}
