\name{model.matrix.CARBayesST}
\alias{model.matrix.CARBayesST}


\title{
    Extract the Model (design) matrix from a model.
}
\description{
    This function takes a CARBayesST object and returns the design matrix.
}

\usage{
    \method{model.matrix}{CARBayesST}(object, \dots)
}



\arguments{
    \item{object}{
        A CARBayesST fitted model object.
    }
    \item{\dots}{
        Ignored.
    }
}
\author{
    Duncan Lee
}
