\name{logLik.CARBayesST}
\alias{logLik.CARBayesST}


\title{
    Extract the estimated loglikelihood from a fitted model.
}
\description{
    This function takes a CARBayesST object and returns the estimated loglikelihood
     based on posterior medians.
}

\usage{
    \method{logLik}{CARBayesST}(object, \dots)
}



\arguments{
    \item{object}{
        A CARBayesST fitted model object.
    }
    \item{\dots}{
        Ignored.
    }
}
\author{
    Duncan Lee
}
