\name{print.carbayesST}
\alias{print.carbayesST}


\title{
Print a summary of the fitted model to the screen.
}
\description{
This function takes a carbayesST object and returns a summary of the fitted model.
}

\usage{
\method{print}{carbayesST}(x, \dots)
}



\arguments{
  \item{x}{
A carbayesST fitted model object
}
  \item{\dots}{
Ignored
}
}
\author{
Duncan Lee
}

\examples{
## See the examples in the main help file for CARBayes.
}